[#162167584] Fixes activation of verifica when resetting payment state
also makes withErrorModal to use a simple overlay screen
instead of a real modal
